Re-enable opcache.validate_timestamps
This change disables a non-standard optimization that was applied automatically by Bref. The main reason it was applied was that the code source is read-only on Lambda.

However, that is not 100% true: code written to `/tmp` is read/write (for example when using `/tmp` as a cache directory, or for a compiled container, etc.). The `opcache.validate_timestamps=0` would cause bugs very hard to debug.

Additionally, switching to `opcache.validate_timestamps=1` will allow code in development to be reloaded when changed without having to restart containers.

This could cause a very small performance regression, but I consider that less important than fixing an actual bug.

Users are welcome to re-enable that option in `php.ini`:

```
opcache.validate_timestamps=0
```